requirements.txt
apache-beam==2.48.0 ; python_version >= "3.9" and python_version < "4.0"
apache-flink-libraries==1.18.0 ; python_version >= "3.9" and python_version < "4.0"
apache-flink==1.18.0 ; python_version >= "3.9" and python_version < "4.0"
asyncio==3.4.3 ; python_version >= "3.9" and python_version < "4.0"
avro-python3==1.10.2 ; python_version >= "3.9" and python_version < "4.0"
black==23.11.0 ; python_version >= "3.9" and python_version < "4.0"
certifi==2023.11.17 ; python_version >= "3.9" and python_version < "4.0"
cffi==1.16.0 ; python_version >= "3.9" and python_version < "4.0" and platform_python_implementation == "PyPy"
charset-normalizer==3.3.2 ; python_version >= "3.9" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.9" and python_version < "4.0"
cloudpickle==2.2.1 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
crcmod==1.7 ; python_version >= "3.9" and python_version < "4.0"
dill==0.3.1.1 ; python_version >= "3.9" and python_version < "4.0"
dnspython==2.4.2 ; python_version >= "3.9" and python_version < "4.0"
docopt==0.6.2 ; python_version >= "3.9" and python_version < "4.0"
fastavro==1.9.0 ; python_version >= "3.9" and python_version < "4.0"
fasteners==0.19 ; python_version >= "3.9" and python_version < "4.0"
find-libpython==0.3.1 ; python_version >= "3.9" and python_version < "4.0" and platform_system != "Windows"
grpcio-tools==1.59.3 ; python_version >= "3.9" and python_version < "4.0"
grpcio==1.59.3 ; python_version >= "3.9" and python_version < "4.0"
hdfs==2.7.3 ; python_version >= "3.9" and python_version < "4.0"
httplib2==0.22.0 ; python_version >= "3.9" and python_version < "4.0"
idna==3.6 ; python_version >= "3.9" and python_version < "4.0"
loguru==0.7.2 ; python_version >= "3.9" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.9" and python_version < "4.0"
numpy==1.24.4 ; python_version >= "3.9" and python_version < "4.0"
objsize==0.6.1 ; python_version >= "3.9" and python_version < "4.0"
orjson==3.9.10 ; python_version >= "3.9" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.9" and python_version < "4.0"
pandas==2.1.3 ; python_version >= "3.9" and python_version < "4.0"
pathspec==0.11.2 ; python_version >= "3.9" and python_version < "4.0"
pemja==0.3.0 ; python_version >= "3.9" and python_version < "4.0" and platform_system != "Windows"
platformdirs==4.0.0 ; python_version >= "3.9" and python_version < "4.0"
proto-plus==1.22.3 ; python_version >= "3.9" and python_version < "4.0"
protobuf==4.23.4 ; python_version >= "3.9" and python_version < "4.0"
py4j==0.10.9.7 ; python_version >= "3.9" and python_version < "4.0"
pyarrow==11.0.0 ; python_version >= "3.9" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.9" and python_version < "4.0" and platform_python_implementation == "PyPy"
pydot==1.4.2 ; python_version >= "3.9" and python_version < "4.0"
pymongo==4.6.0 ; python_version >= "3.9" and python_version < "4.0"
pyparsing==3.1.1 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "4.0"
pytz==2023.3.post1 ; python_version >= "3.9" and python_version < "4.0"
regex==2023.10.3 ; python_version >= "3.9" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.9" and python_version < "4.0"
setuptools==69.0.2 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.9" and python_version < "3.11"
typing-extensions==4.8.0 ; python_version >= "3.9" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.9" and python_version < "4.0"
urllib3==2.1.0 ; python_version >= "3.9" and python_version < "4.0"
win32-setctime==1.1.0 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "win32"
zstandard==0.22.0 ; python_version >= "3.9" and python_version < "4.0"
asyncio==3.4.3
kafka-python==2.0.2
asyncio==3.4.3
websockets==13.0.1
looseversion==1.3.0
rethinkdb==2.4.10.post1
six==1.16.0
websockets==13.0.1
cassandra-driver==3.29.2
click==8.1.7
Cython==3.0.11
geomet==0.2.1.post1
happybase==1.2.0
ply==3.11
six==1.16.0
thriftpy2==0.5.2
